requirements.txt
flask==0.8
flask-login
flask-wtf
blinker
pyyaml
mongoengine
twilio
nose
fabric